Remove deprecated modules and integrate high-performance components into agent (#10)
